Fix warning introduced  checking U-Boot variable length
Commit 87565c2693acfc049102ba07268286ef1d5be6ef  introduces
these warnings:

parser/parser.c: In function 'check_field_string':
include/util.h:60:11: warning: format '%lu' expects argument of type
'long unsigned int', but argument 8 has type 'unsigned int' [-Wformat=]
           "ERROR %s : %s : %d : " format, \
           ^
include/util.h:82:2: note: in expansion of macro 'swupdate_notify'
  swupdate_notify(RUN, format, WARNLEVEL, ## arg)
  ^
parser/parser.c:81:3: note: in expansion of macro 'WARN'
   WARN("Configuration Key '%s...' is > %lu chars, cropping it.\n",
   ^

CC: Christian Storm <christian.storm@tngtech.com>

+++ parser/parser.c
@@ -78,8 +78,8 @@ static void check_field_string(const char *src, char *dst, const size_t max_len)
 	size_t act_len = strnlen(src, SWUPDATE_GENERAL_STRING_SIZE);
 	if (act_len > max_len) {
 		((char*)dst)[max_len-1] = '\0';
-		WARN("Configuration Key '%s...' is > %lu chars, cropping it.\n",
-			(char*)dst, max_len-1);
+		WARN("Configuration Key '%s...' is > %u chars, cropping it.\n",
+			(char*)dst, (unsigned int)max_len - 1);
 	}
 	if (act_len == 0) {
 		WARN("Configuration Key is empty!\n");
